requirements.txt
torch>=2.0.1
torchvision>=0.15.2
numpy>=1.23.5
pytorch-lightning[extra]==2.0.2
torchmetrics==0.11.0
timm>=0.6.12
argparse>=1.4.0
mlflow>=2.3.2
albumentations>=1.3.0
randaugment>=1.0.2
pillow>=9.3.0
efficientnet-pytorch>=0.7.1
madgrad>=1.2
git+https://github.com/MIC-DKFZ/dynamic-network-architectures.git
rich>=12.6.0
matplotlib>=3.6.2